+++ RedPandaIDE/cpprefacter.cpp
@@ -16,6 +16,8 @@ CppRefacter::CppRefacter(QObject *parent) : QObject(parent)
 
 bool CppRefacter::findOccurence(Editor *editor, const BufferCoord &pos)
 {
+    if (!editor->parser())
+        return false;
     if (!editor->parser()->freeze())
         return false;
     auto action = finally([&editor]{
@@ -35,36 +37,50 @@ bool CppRefacter::findOccurence(Editor *editor, const BufferCoord &pos)
 
     std::shared_ptr<Project> project = pMainWindow->project();
     if (editor->inProject() && project) {
-        PSearchResults results = pMainWindow->searchResultModel()->addSearchResults(
-                    statement->fullName,
-                    SearchFileScope::wholeProject
-                    );
-        foreach (const PProjectUnit& unit, project->units()) {
-            if (isCfile(unit->fileName()) || isHfile(unit->fileName())) {
-                PSearchResultTreeItem item = findOccurenceInFile(
-                            unit->fileName(),
-                            statement,
-                            editor->parser());
-                if (item && !(item->results.isEmpty())) {
-                    results->results.append(item);
-                }
-            }
-        }
+        doFindOccurenceInProject(statement,project,editor->parser());
     } else {
-        PSearchResults results = pMainWindow->searchResultModel()->addSearchResults(
-                    statement->fullName,
-                    SearchFileScope::currentFile
-                    );
-        PSearchResultTreeItem item = findOccurenceInFile(
-                    editor->filename(),
-                    statement,
-                    editor->parser());
-        if (item && !(item->results.isEmpty())) {
-            results->results.append(item);
-        }
+        doFindOccurenceInEditor(statement,editor,editor->parser());
     }
     pMainWindow->searchResultModel()->notifySearchResultsUpdated();
     return true;
+}
+
+bool CppRefacter::findOccurence(const QString &statementFullname, SearchFileScope scope)
+{
+    PCppParser parser;
+    Editor * editor;
+    std::shared_ptr<Project> project;
+    if (scope == SearchFileScope::currentFile) {
+        editor = pMainWindow->editorList()->getEditor();
+        if (!editor)
+            return false;
+        parser = editor->parser();
+    } else if (scope == SearchFileScope::wholeProject) {
+        project = pMainWindow->project();
+        if (!project)
+            return false;
+        parser = project->cppParser();
+    }
+    if (!parser)
+        return false;
+    {
+        parser->freeze();
+        auto action = finally([&parser]{
+            parser->unFreeze();
+        });
+        PStatement statement = parser->findStatement(statementFullname);
+        // definition of the symbol not found
+        if (!statement)
+            return false;
+
+        if (scope == SearchFileScope::wholeProject) {
+            doFindOccurenceInProject(statement,project,parser);
+        } else if (scope == SearchFileScope::currentFile) {
+            doFindOccurenceInEditor(statement, editor,parser);
+        }
+        pMainWindow->searchResultModel()->notifySearchResultsUpdated();
+        return true;
+    }
 }
 
 static QString fullParentName(PStatement statement) {
@@ -115,6 +131,42 @@ void CppRefacter::renameSymbol(Editor *editor, const BufferCoord &pos, const QSt
         return;
     }
     renameSymbolInFile(editor->filename(),oldStatement,newWord, editor->parser());
+}
+
+void CppRefacter::doFindOccurenceInEditor(PStatement statement , Editor *editor, const PCppParser &parser)
+{
+    PSearchResults results = pMainWindow->searchResultModel()->addSearchResults(
+                statement->command,
+                statement->fullName,
+                SearchFileScope::currentFile
+                );
+    PSearchResultTreeItem item = findOccurenceInFile(
+                editor->filename(),
+                statement,
+                parser);
+    if (item && !(item->results.isEmpty())) {
+        results->results.append(item);
+    }
+}
+
+void CppRefacter::doFindOccurenceInProject(PStatement statement, std::shared_ptr<Project> project, const PCppParser &parser)
+{
+    PSearchResults results = pMainWindow->searchResultModel()->addSearchResults(
+                statement->command,
+                statement->fullName,
+                SearchFileScope::wholeProject
+                );
+    foreach (const PProjectUnit& unit, project->units()) {
+        if (isCfile(unit->fileName()) || isHfile(unit->fileName())) {
+            PSearchResultTreeItem item = findOccurenceInFile(
+                        unit->fileName(),
+                        statement,
+                        parser);
+            if (item && !(item->results.isEmpty())) {
+                results->results.append(item);
+            }
+        }
+    }
 }
 
 PSearchResultTreeItem CppRefacter::findOccurenceInFile(