+++ RedPandaIDE/cpprefacter.h
@@ -8,6 +8,7 @@
 
 class Editor;
 class BufferCoord;
+class Project;
 class CppRefacter : public QObject
 {
     Q_OBJECT
@@ -15,10 +16,13 @@ public:
     explicit CppRefacter(QObject *parent = nullptr);
 
     bool findOccurence(Editor * editor, const BufferCoord& pos);
+    bool findOccurence(const QString& statementFullname, SearchFileScope scope);
 
     void renameSymbol(Editor* editor, const BufferCoord& pos, const QString& word, const QString& newWord);
 signals:
 private:
+    void doFindOccurenceInEditor(PStatement statement, Editor* editor, const PCppParser& parser);
+    void doFindOccurenceInProject(PStatement statement, std::shared_ptr<Project> project, const PCppParser& parser);
     PSearchResultTreeItem findOccurenceInFile(
             const QString& filename,
             const PStatement& statement,