+++ RedPandaIDE/mainwindow.h
@@ -332,7 +332,7 @@ private slots:
 
     void on_cbSearchHistory_currentIndexChanged(int index);
 
-    void on_btnSearchAgin_clicked();
+    void on_btnSearchAgain_clicked();
     void on_actionRemove_Watch_triggered();
 
     void on_actionRemove_All_Watches_triggered();