+++ RedPandaIDE/parser/cppparser.h
@@ -36,6 +36,7 @@ public:
                             int line);
     int findLastOperator(const QString& phrase) const;
     PStatementList findNamespace(const QString& name); // return a list of PSTATEMENTS (of the namespace)
+    PStatement findStatement(const QString& fullname);
     PStatement findStatementOf(const QString& fileName,
                                const QString& phrase,
                                int line);